+++ tools/uva/server.h
@@ -7,7 +7,7 @@ namespace corelab {
     void* ClientRoutine(void*);
 
     /* These two function do nothing. Everybody are client */
-    extern "C" void uva_server_load(void *addr, uint64_t len);
-    extern "C" void uva_server_store(void *addr, uint64_t len, void *data); 
+    extern "C" void uva_server_load(void *addr, size_t len);
+    extern "C" void uva_server_store(void *addr, size_t len, void *data); 
   }
 }
